// overlay/src/client/message_schema.ts
export class ColorDto {
    public r: number;
    public g: number;
    public b: number;
    constructor(params: {
        r: number,
        g: number,
        b: number,
    }) {
        this.r = params.r;
        this.g = params.g;
        this.b = params.b;
    }
    public toJsonValue(): JsonObject{
        return {
            r: this.r,
            g: this.g,
            b: this.b,
        }
    }
    public static fromJsonValue(value: JsonValue): ColorDto | Error{
        return parse_as_ColorDto(value)    }
}
// ...
export class DataSourceDto {
    public url: UrlDto;
    public interval: Interval5DDto;
    public tile_shape: Shape5DDto;
    public spatial_resolution: [number,number,number];
    constructor(params: {
        url: UrlDto,
        interval: Interval5DDto,
        tile_shape: Shape5DDto,
        spatial_resolution: [number,number,number],
    }) {
        this.url = params.url;
        this.interval = params.interval;
        this.tile_shape = params.tile_shape;
        this.spatial_resolution = params.spatial_resolution;
    }
    public toJsonValue(): JsonObject{
        return {
            url: this.url.toJsonValue(),
            interval: this.interval.toJsonValue(),
            tile_shape: this.tile_shape.toJsonValue(),
            spatial_resolution: [this.spatial_resolution[0],
this.spatial_resolution[1],
this.spatial_resolution[2]],
        }
    }
    public static fromJsonValue(value: JsonValue): DataSourceDto | Error{
        return parse_as_DataSourceDto(value)    }
}
// ...
export class PixelAnnotationDto {
    public raw_data: DataSourceDto;
    public points: Array<[number,number,number]>;
    constructor(params: {
        raw_data: DataSourceDto,
        points: Array<[number,number,number]>,
    }) {
        this.raw_data = params.raw_data;
        this.points = params.points;
    }
    public toJsonValue(): JsonObject{
        return {
            raw_data: this.raw_data.toJsonValue(),
            points: this.points.map(item => [item[0],
item[1],
item[2]]),
        }
    }
    public static fromJsonValue(value: JsonValue): PixelAnnotationDto | Error{
        return parse_as_PixelAnnotationDto(value)    }
}
// ...
export class RecolorLabelParams {
    public label_name: string;
    public new_color: ColorDto;
    constructor(params: {
        label_name: string,
        new_color: ColorDto,
    }) {
        this.label_name = params.label_name;
        this.new_color = params.new_color;
    }
    public toJsonValue(): JsonObject{
        return {
            label_name: this.label_name,
            new_color: this.new_color.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): RecolorLabelParams | Error{
        return parse_as_RecolorLabelParams(value)    }
}
// ...
export class RenameLabelParams {
    public old_name: string;
    public new_name: string;
    constructor(params: {
        old_name: string,
        new_name: string,
    }) {
        this.old_name = params.old_name;
        this.new_name = params.new_name;
    }
    public toJsonValue(): JsonObject{
        return {
            old_name: this.old_name,
            new_name: this.new_name,
        }
    }
    public static fromJsonValue(value: JsonValue): RenameLabelParams | Error{
        return parse_as_RenameLabelParams(value)    }
}
// ...
export class CreateLabelParams {
    public label_name: string;
    public color: ColorDto;
    constructor(params: {
        label_name: string,
        color: ColorDto,
    }) {
        this.label_name = params.label_name;
        this.color = params.color;
    }
    public toJsonValue(): JsonObject{
        return {
            label_name: this.label_name,
            color: this.color.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): CreateLabelParams | Error{
        return parse_as_CreateLabelParams(value)    }
}
// ...
export class RemoveLabelParams {
    public label_name: string;
    constructor(params: {
        label_name: string,
    }) {
        this.label_name = params.label_name;
    }
    public toJsonValue(): JsonObject{
        return {
            label_name: this.label_name,
        }
    }
    public static fromJsonValue(value: JsonValue): RemoveLabelParams | Error{
        return parse_as_RemoveLabelParams(value)    }
}
// ...
export class AddPixelAnnotationParams {
    public label_name: string;
    public pixel_annotation: PixelAnnotationDto;
    constructor(params: {
        label_name: string,
        pixel_annotation: PixelAnnotationDto,
    }) {
        this.label_name = params.label_name;
        this.pixel_annotation = params.pixel_annotation;
    }
    public toJsonValue(): JsonObject{
        return {
            label_name: this.label_name,
            pixel_annotation: this.pixel_annotation.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): AddPixelAnnotationParams | Error{
        return parse_as_AddPixelAnnotationParams(value)    }
}
// ...
export class RemovePixelAnnotationParams {
    public label_name: string;
    public pixel_annotation: PixelAnnotationDto;
    constructor(params: {
        label_name: string,
        pixel_annotation: PixelAnnotationDto,
    }) {
        this.label_name = params.label_name;
        this.pixel_annotation = params.pixel_annotation;
    }
    public toJsonValue(): JsonObject{
        return {
            label_name: this.label_name,
            pixel_annotation: this.pixel_annotation.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): RemovePixelAnnotationParams | Error{
        return parse_as_RemovePixelAnnotationParams(value)    }
}